Crate capsules

Reexports

extern crate core as core;

extern crate kernel;

use core::prelude::v1::*;

Modules

adc
Provides userspace applications with the ability to sample analog signals.

button
Provides capsule driver for controlling buttons on a board.

console
Provides userspace with the ability to print text via a serial interface.

crc
Provides userspace access to a CRC unit.

dac
Provides a DAC interface for userspace.

fm25cl
Driver for the FM25CL FRAM chip.

fxos8700cq
Driver for the FXOS8700CQ accelerometer.

gpio
Provides userspace applications with access to GPIO pins.

gpio_async
Provides userspace applications with a driver interface to asynchronous GPIO pins.

i2c_master_slave_driver
Provides both an I2C Master and I2C Slave interface to userspace.

isl29035
Driver for the ISL29035 digital light sensor.

led
Provides userspace access to LEDs on a board.

lps25hb
Driver for the ST LPS25HB pressure sensor.

ltc294x
Driver for the LTC294X line of coulomb counters.

max17205
Driver for the Maxim MAX17205 fuel gauge.

mcp23008
Driver for the Microchip MCP23008 I2C GPIO extender.

ninedof
Provides userspace with virtualized access to 9DOF sensors.

nonvolatile_storage_driver
This provides kernel and userspace access to nonvolatile memory.

nonvolatile_to_pages
Map arbitrary nonvolatile reads and writes to page operations.

nrf51822_serialization
Provides userspace with the UART API that the nRF51822 serialization library requires.

pca9544a
Driver for the PCA9544A I2C Selector.

radio
Provides userspace applications with the ability to send and receive 802.15.4 packets.

rf233
Driver for sending 802.15.4 packets with an Atmel RF233.

rf233_const
Support for the RF233 capsule.

rng
Provides a simple driver for userspace applications to request randomness.

sdcard
Provides driver for accessing an SD Card and a userspace Driver.

si7021
Driver for the Silicon Labs SI7021 temperature/humidity sensor.

spi
Provides userspace applications with the ability to communicate over the SPI bus.

symmetric_encryption
Provides userspace applications with the ability to encrypt and decrypt messages.

temp_nrf51dk
Provides userspace with access to the NRF51DK onboard temperature sensor.

timer
Provides userspace applications with a timer API.

tmp006
Driver for the TI TMP006 infrared thermopile contactless temperature sensor.

tsl2561
Driver for the Taos TSL2561 light sensor.

virtual_alarm
Virtualize the Alarm interface to enable multiple users of an underlying alarm hardware peripheral.

virtual_flash
Virtualize writing flash.

virtual_i2c
Virtualize an I2C master bus.

virtual_spi
Virtualize a SPI master bus to enable multiple users of the SPI bus.