Struct capsules::console::Console

pub struct Console<'a, U: UART + 'a> {
    uart: &'a U,
    apps: Container<App>,
    in_progress: Cell<Option<AppId>>,
    tx_buffer: TakeCell<'static, [u8]>,
    baud_rate: u32,
}

Fields
uart: &'a U apps: Container<App> in_progress: Cell<Option<AppId>> tx_buffer: TakeCell<'static, [u8]> baud_rate: u32
Methods

impl<'a, U: UART> Console<'a, U>

fn new(
uart: &'a U,
baud_rate: u32,
tx_buffer: &'static mut [u8],
container: Container<App>
) -> Console<'a, U>

fn initialize(&self)

fn send_new(
&self,
app_id: AppId,
app: &mut App,
callback: Callback
) -> ReturnCode

Internal helper function for setting up a new send transaction

fn send_continue(
&self,
app_id: AppId,
app: &mut App
) -> Result<bool, ReturnCode>

Internal helper function for continuing a previously set up transaction Returns true if this send is still active, or false if it has completed

fn send(&self, app_id: AppId, app: &mut App, slice: AppSlice<Shared, u8>)

Internal helper function for sending data for an existing transaction. Cannot fail. If can't send now, it will schedule for sending later.