Struct capsules::crc::Crc

pub struct Crc<'a, C: CRC + 'a> {
    crc_unit: &'a C,
    apps: Container<App>,
    serving_app: Cell<Option<AppId>>,
}

Struct that holds the state of the CRC driver and implements the Driver trait for use by processes through the system call interface.

Fields
crc_unit: &'a C apps: Container<App> serving_app: Cell<Option<AppId>>
Methods

impl<'a, C: CRC> Crc<'a, C>

fn new(crc_unit: &'a C, apps: Container<App>) -> Crc<'a, C>

Create a Crc driver

The argument crc_unit must implement the abstract CRC hardware interface. The argument apps should be an empty kernel Container, and will be used to track application requests.

Example

capsules::crc::Crc::new(&sam4l::crccu::CRCCU, kernel::Container::create()),

fn serve_waiting_apps(&self)