Module capsules::led

Provides userspace access to LEDs on a board.

This allows for much more cross platform controlling of LEDs without having to know which of the GPIO pins exposed across the syscall interface are LEDs.

Usage

let led_pins = static_init!(
    [(&'static sam4l::gpio::GPIOPin, capsules::led::ActivationMode); 3],
    [(&sam4l::gpio::PA[13], capsules::led::ActivationMode::ActiveLow),   // Red
     (&sam4l::gpio::PA[15], capsules::led::ActivationMode::ActiveLow),   // Green
     (&sam4l::gpio::PA[14], capsules::led::ActivationMode::ActiveLow)]); // Blue
let led = static_init!(
    capsules::led::LED<'static, sam4l::gpio::GPIOPin>,
    capsules::led::LED::new(led_pins));

Reexports

use kernel::AppId;

use kernel::Driver;

use kernel::ReturnCode;

use kernel::hil;

Structs

LED

Enums

ActivationMode
Whether the LEDs are active high or active low on this platform.