Struct capsules::mcp23008::MCP23008

pub struct MCP23008<'a> {
    i2c: &'a I2CDevice,
    state: Cell<State>,
    buffer: TakeCell<'static, [u8]>,
    interrupt_pin: Option<&'static Pin>,
    interrupt_settings: Cell<u32>,
    identifier: Cell<usize>,
    client: Cell<Option<&'static Client>>,
}

Fields
i2c: &'a I2CDevice state: Cell<State> buffer: TakeCell<'static, [u8]> interrupt_pin: Option<&'static Pin> interrupt_settings: Cell<u32> identifier: Cell<usize> client: Cell<Option<&'static Client>>
Methods

impl<'a> MCP23008<'a>

fn new(
i2c: &'a I2CDevice,
interrupt_pin: Option<&'static Pin>,
buffer: &'static mut [u8]
) -> MCP23008<'a>

fn set_client<C: Client>(&self, client: &'static C)

Set the client of this MCP23008 when commands finish or interrupts occur. The identifier is simply passed back with the callback so that the upper layer can keep track of which device triggered.

fn enable_host_interrupt(&self) -> ReturnCode

fn set_direction(&self, pin_number: u8, direction: Direction) -> ReturnCode

fn configure_pullup(&self, pin_number: u8, enabled: bool) -> ReturnCode

Set the pull-up on the pin also configure it to be an input.

fn set_pin(&self, pin_number: u8, value: PinState) -> ReturnCode

fn toggle_pin(&self, pin_number: u8) -> ReturnCode

fn read_pin(&self, pin_number: u8) -> ReturnCode

fn enable_interrupt_pin(
&self,
pin_number: u8,
direction: InterruptMode
) -> ReturnCode

fn disable_interrupt_pin(&self, pin_number: u8) -> ReturnCode

fn save_pin_interrupt_state(
&self,
pin_number: u8,
enabled: bool,
direction: InterruptMode
)

Helper function for keeping track of which interrupts are currently enabled.

fn remove_pin_interrupt_state(&self, pin_number: u8)

fn get_pin_interrupt_enabled_state(&self) -> u8

Create an 8 bit bitmask of which interrupts are enabled.

fn check_pin_interrupt_enabled(&self, pin_number: u8) -> bool

fn get_pin_interrupt_direction(&self, pin_number: u8) -> InterruptMode