Struct capsules::nonvolatile_to_pages::NonvolatileToPages

pub struct NonvolatileToPages<'a, F: Flash + 'static> {
    driver: &'a F,
    client: Cell<Option<&'static NonvolatileStorageClient>>,
    pagebuffer: TakeCell<'static, F::Page>,
    state: Cell<State>,
    buffer: TakeCell<'static, [u8]>,
    address: Cell<usize>,
    length: Cell<usize>,
    remaining_length: Cell<usize>,
    buffer_index: Cell<usize>,
}

Fields
driver: &'a F

The module providing a Flash interface.

client: Cell<Option<&'static NonvolatileStorageClient>>

Callback to the user of this capsule.

pagebuffer: TakeCell<'static, F::Page>

Buffer correctly sized for the underlying flash page size.

state: Cell<State>

Current state of this capsule.

buffer: TakeCell<'static, [u8]>

Temporary holding place for the user's buffer.

address: Cell<usize>

Absolute address of where we are reading or writing. This gets updated as the operation proceeds across pages.

length: Cell<usize>

Total length to read or write. We need to store this to return it to the client.

remaining_length: Cell<usize>

How many bytes are left to read or write.

buffer_index: Cell<usize>

Where we are in the user buffer.

Methods

impl<'a, F: Flash + 'a> NonvolatileToPages<'a, F>

fn new(driver: &'a F, buffer: &'static mut F::Page) -> NonvolatileToPages<'a, F>