Module capsules::rf233

Driver for sending 802.15.4 packets with an Atmel RF233.

This implementation is completely non-blocking. This means that the state machine is somewhat complex, as it must interleave interrupt handling with requests and radio state management. See the SPI read_write_done handler for details.

To do items:

Support TX power control

Support channel selection

Support link-layer acknowledgements

Support power management (turning radio off)

Reexports

use core::cell::Cell;

use kernel::ReturnCode;

use kernel::common::take_cell::TakeCell;

use kernel::hil::gpio;

use kernel::hil::radio;

use kernel::hil::spi;

use rf233_const::*;

Structs

RF233

Enums

InternalState

Constants

INTERRUPT_ID

Functions

interrupt_included

power_to_setting

setting_to_power