Struct capsules::sdcard::SDCardDriver

pub struct SDCardDriver<'a, A: Alarm + 'a> {
    sdcard: &'a SDCard<'a, A>,
    app: MapCell<App>,
    kernel_buf: TakeCell<'static, [u8]>,
}

Application driver for SD Card capsule, layers on top of SD Card capsule This is used if the SDCard is going to be attached directly to userspace syscalls. SDCardDriver can be ignored if another capsule is going to build off of the SDCard instead

Fields
sdcard: &'a SDCard<'a, A> app: MapCell<App> kernel_buf: TakeCell<'static, [u8]>
Methods

impl<'a, A: Alarm + 'a> SDCardDriver<'a, A>

Functions for SDCardDriver

fn new(
sdcard: &'a SDCard<'a, A>,
kernel_buf: &'static mut [u8; 512]
) -> SDCardDriver<'a, A>

Create new SD card userland interface

sdcard - SDCard interface to provide application access to kernel_buf - buffer used to hold SD card blocks, must be at least 512 bytes in length