Module capsules::tmp006

Driver for the TI TMP006 infrared thermopile contactless temperature sensor.

http://www.ti.com/product/TMP006

The TMP006 and TMP006B are fully integrated MEMs thermopile sensors that measure the temperature of an object without having to be in direct contact. The thermopile absorbs passive infrared energy from an object at wavelengths between 4 um to 16 um within the end-user defined field of view.

Reexports

use core::cell::Cell;

use kernel::AppId;

use kernel::Callback;

use kernel::Driver;

use kernel::ReturnCode;

use kernel::common::math::sqrtf32;

use kernel::common::math::get_errno;

use kernel::common::take_cell::TakeCell;

use kernel::hil::gpio::Pin;

use kernel::hil::gpio::InterruptMode;

use kernel::hil::gpio::Client;

use kernel::hil::i2c;

Structs

TMP006

Enums

ProtocolState
States of the I2C protocol with the TMP006. There are three sequences:

Registers

Constants

A_1

A_2

B_0

B_1

B_2

C_2

C_TO_K

DEFAULT_SAMPLING_RATE

K_TO_C

MAX_SAMPLING_RATE

NV_TO_V

S_0

T_DIE_CONVERT

T_REF

V_OBJ_CONVERT

Statics

BUFFER

Functions

calculate_temperature

Type Definitions

SensorVoltage