Module capsules::tsl2561

Driver for the Taos TSL2561 light sensor.

http://www.digikey.com/product-detail/en/ams-taos-usa-inc/TSL2561FN/TSL2561-FNCT-ND/3095298

The TSL2560 and TSL2561 are light-to-digital converters that transform light intensity to a digital signal output capable of direct I2C interface. Each device combines one broadband photodiode (visible plus infrared) and one infrared-responding photodiodeon a single CMOS integrated circuit capable of providing a near-photopic response over an effective 20-bit dynamic range (16-bit resolution). Two integrating ADCs convert the photodiode currents to a digital output that represents the irradiance measured on each channel. This digital output can be input to a microprocessor where illuminance (ambient light level) in lux is derived using an empirical formula to approximate the human eye response.

Reexports

use core::cell::Cell;

use kernel::AppId;

use kernel::Callback;

use kernel::Driver;

use kernel::ReturnCode;

use kernel::common::take_cell::TakeCell;

use kernel::hil::gpio;

use kernel::hil::i2c;

Structs

TSL2561

Enums

Registers

State

Constants

B1T

B2T

B3T

B4T

B5T

B6T

B7T

B8T

CHSCALE_TINT0

CHSCALE_TINT1

CH_SCALE

COMMAND_REG
Command register defines

INTEGRATE_TIME_101_MS
Timing_Reg defines

INTERRUPT_CONTROL_LEVEL

INTERRUPT_ON_ADC_DONE

K1T

K2T

K3T

K4T

K5T

K6T

K7T

K8T

LOW_GAIN_MODE

LUX_SCALE

M1T

M2T

M3T

M4T

M5T

M6T

M7T

M8T

POWER_OFF

POWER_ON
Control_Reg defines

RATIO_SCALE

WORD_PROTOCOL

Statics

BUFFER