Struct capsules::virtual_flash::FlashUser

pub struct FlashUser<'a, F: Flash + 'static> {
    mux: &'a MuxFlash<'a, F>,
    buffer: TakeCell<'static, F::Page>,
    operation: Cell<Op>,
    next: ListLink<'a, FlashUser<'a, F>>,
    client: Cell<Option<&'a Client<FlashUser<'a, F>>>>,
}

Keep state for each flash user. All uses of the virtualized flash interface need to create one of these to be a user of the flash. The new() function handles most of the work, a user only has to pass in a reference to the MuxFlash object.

Fields
mux: &'a MuxFlash<'a, F> buffer: TakeCell<'static, F::Page> operation: Cell<Op> next: ListLink<'a, FlashUser<'a, F>> client: Cell<Option<&'a Client<FlashUser<'a, F>>>>
Methods

impl<'a, F: Flash + 'a> FlashUser<'a, F>

const fn new(mux: &'a MuxFlash<'a, F>) -> FlashUser<'a, F>