Struct capsules::virtual_flash::MuxFlash

pub struct MuxFlash<'a, F: Flash + 'static> {
    flash: &'a F,
    users: List<'a, FlashUser<'a, F>>,
    inflight: Cell<Option<&'a FlashUser<'a, F>>>,
}

Handle keeping a list of active users of flash hardware and serialize their requests. After each completed request the list is checked to see if there is another flash user with an outstanding read, write, or erase request.

Fields
flash: &'a F users: List<'a, FlashUser<'a, F>> inflight: Cell<Option<&'a FlashUser<'a, F>>>
Methods

impl<'a, F: Flash + 'a> MuxFlash<'a, F>

const fn new(flash: &'a F) -> MuxFlash<'a, F>

fn do_next_op(&self)

Scan the list of users and find the first user that has a pending request, then issue that request to the flash hardware.