Module capsules::virtual_spi

Virtualize a SPI master bus to enable multiple users of the SPI bus.

Reexports

use core::cell::Cell;

use kernel::ReturnCode;

use kernel::common::List;

use kernel::common::ListLink;

use kernel::common::ListNode;

use kernel::common::take_cell::TakeCell;

use kernel::hil;

Structs

MuxSpiMaster
The Mux struct manages multiple Spi clients. Each client may have at most one outstanding Spi request.

VirtualSpiMasterDevice

VirtualSpiSlaveDevice

Enums

Op